+++ autoload.php
@@ -2,6 +2,9 @@
 
 require_once __DIR__ . '/nspl/nspl.php';
 require_once __DIR__ . '/nspl/a.php';
+require_once __DIR__ . '/nspl/a/ChainableSequence.php';
+require_once __DIR__ . '/nspl/a/ChainableArray.php';
+require_once __DIR__ . '/nspl/a/lazy/LazyChainableSequence.php';
 
 if ((PHP_MAJOR_VERSION >= 7) || (PHP_MAJOR_VERSION >= 5 && PHP_MINOR_VERSION >= 4)) {
     require_once __DIR__ . '/nspl/a/lazy.php';